Finalize mongo app

+++ Rest_API/src/main/java/com/clamaud/restAPI/repository/UserRepository.java
@@ -1,5 +1,6 @@
 package com.clamaud.restAPI.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -12,7 +13,7 @@ import com.clamaud.restAPI.domain.jpa.User;
  * Repository Spring DATA
  */
 
-public interface UserRepository extends MongoRepository<User, Long>{
+public interface UserRepository extends MongoRepository<User, BigInteger>{
 
 	List<User> findByLastName(String lastName);
